Updated generation skeleton to use Arrays of ints
Generation code does not use a Tilemap internally anymore.  It should
initialize a 2d array of ints, which are the integer tile types that
represent the tiles.  These tile types do not have a 1-1 matching with
tiles, as there can be tile types like GRASS and DIRT, which both map to
the GRASS tile.

Also code was cleaned up in terms of whitespace.